Move forward index builder implementation to cpp
Moves the implementation to a separate compilation unit for faster
compilation. It also includes some minor changes to eliminate most of
the compilation warnings.
