+++ external/CMakeLists.txt
@@ -30,7 +30,7 @@ add_library(MaskedVByte STATIC ${CMAKE_CURRENT_SOURCE_DIR}/MaskedVByte/src/varin
 
 # Add QMX
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/QMX EXCLUDE_FROM_ALL)
-target_compile_options(QMX INTERFACE -Wno-implicit-fallthrough)
+target_compile_options(QMX INTERFACE -Wno-unused-parameter -Wno-implicit-fallthrough)
 
 # Add mio
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/mio EXCLUDE_FROM_ALL)
