+++ include/pisa/query/query_stemmer.hpp
@@ -3,15 +3,18 @@
 #include <sstream>
 #include <string>
 
+#include <boost/algorithm/string/join.hpp>
+
 #include "query/queries.hpp"
 #include "query/term_processor.hpp"
 #include "tokenizer.hpp"
-#include <boost/algorithm/string/join.hpp>
+
 namespace pisa {
+
 class QueryStemmer {
   public:
     explicit QueryStemmer(std::optional<std::string> const& stemmer_name)
-        : m_stemmer(term_processor_builder(stemmer_name)())
+        : m_stemmer(term_transformer_builder(stemmer_name)())
     {}
     std::string operator()(std::string const& query_string)
     {
@@ -30,6 +33,7 @@ class QueryStemmer {
         return tokenized_query.str();
     }
 
-    Stemmer_t m_stemmer;
+    TermTransformer m_stemmer;
 };
+
 }  // namespace pisa
