+++ include/pisa/scorer/quantized.hpp
@@ -11,10 +11,9 @@ namespace pisa {
 template <typename Wand>
 struct quantized: public index_scorer<Wand> {
     using index_scorer<Wand>::index_scorer;
-    term_scorer_t term_scorer(uint64_t term_id) const
+    term_scorer_t term_scorer([[maybe_unused]] uint64_t term_id) const
     {
-        auto s = [](uint32_t doc, uint32_t freq) { return freq; };
-        return s;
+        return []([[maybe_unused]] uint32_t doc, uint32_t freq) { return freq; };
     }
 };
